use LONG instead of int for rv in test_rv() The type was wrong on amd64 since the sign was extended
git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/PCSC@2706 0ce88b0d-b2fd-0310-8134-9614164e65ea

+++ src/testpcsc.c
@@ -27,8 +27,8 @@
 #define DONT_PANIC 1
 #define RED_PRINTF_FORMAT "\33[01;31m%s\33[0m\n"
 
-void test_rv(int rv, SCARDCONTEXT hContext, int dont_panic);
-void test_rv(int rv, SCARDCONTEXT hContext, int dont_panic)
+void test_rv(LONG rv, SCARDCONTEXT hContext, int dont_panic);
+void test_rv(LONG rv, SCARDCONTEXT hContext, int dont_panic)
 {
 	if (rv != SCARD_S_SUCCESS)
 	{
